there are two versions checked in to this release now.
The enter-exit-version is the closest to the api talked about with Greig
Has a message for enter a region and one for exit a region.
I didn't quite understand the arguments for his schedule generator so I just adapted mine as I understand that one

The snapshot-version is the same as the above but it only operates on message received every 15 minutes.  It's currently faster than the one above.